feat(T09): refactoring

+++ src/components/TodoLayout/TodoList/components/TodoListHeader/TodoListHeader.tsx
@@ -1,19 +1,17 @@
-import { NotificationType } from "@/common/constants/NotificationConstants.ts";
-import { CurrentTodoFilter, TodoErrorMessages } from "@/common/constants/TodoConstants.ts";
-import { FilteredTodo } from "@/components/TodoLayout/TodoList/components/FilteredTodo/FilteredTodo.tsx";
-import { Button } from "@/components/ui/Button/Button.tsx";
-import { useAppDispatch } from "@/hooks/useAppDispatch.ts";
-import { setNotification } from "@/store/actions/notificationActionCreators.ts";
-import { deleteCompletedTodos, setFiltrationValue } from "@/store/actions/todoActionCreators.ts";
+import { NotificationType, TodoConfirmMessages, TodoCurrentFilter, TodoNotificationMessages } from "@constants";
+import { TodoFiltered } from "../TodoFiltered";
+import { Button } from "@components";
+import { useAppDispatch, useAppSelector, useModalState } from "@hooks";
+import { deleteCompletedTodos, setFiltrationValue, setNotification } from "@store";
 import cn from "classnames";
 import styles from "./TodoListHeader.module.scss";
-import { useAppSelector } from "@/hooks/useAppSelector.ts";
-import { useEffect, useState } from "react";
-import filteredStyles from "@/components/TodoLayout/TodoList/components/FilteredTodo/FilteredTodo.module.scss";
+import { ReactNode, useEffect, useState } from "react";
+import filteredStyles from "../TodoFiltered/TodoFiltered.module.scss";
 
-export const TodoListHeader = () => {
+export const TodoListHeader = (): ReactNode => {
   const { originalTodos } = useAppSelector(state => state.todoReducer);
   const [completedTodo, setCompletedTodo] = useState(0);
+  const setConfirmModalActive = useModalState("confirmModal")[1];
   const dispatch = useAppDispatch();
 
   useEffect(() => {
@@ -21,10 +19,15 @@ export const TodoListHeader = () => {
     setCompletedTodo(completedTodos.length);
   }, [originalTodos]);
 
-  const handleDeleteCompletedTodos = () => {
-    dispatch(setNotification({ title: TodoErrorMessages.CLEAR_COMPLETED, type: NotificationType.SUCCESS }));
-    dispatch(deleteCompletedTodos());
-    dispatch(setFiltrationValue(CurrentTodoFilter.ALL));
+  const handleDeleteCompletedTodos = (): void => {
+    setConfirmModalActive(true, {
+      confirmCallback: () => {
+        dispatch(setNotification({ title: TodoNotificationMessages.DELETE_COMPLETED_TODOS, type: NotificationType.SUCCESS }));
+        dispatch(deleteCompletedTodos());
+        dispatch(setFiltrationValue(TodoCurrentFilter.ALL));
+      },
+      message: TodoConfirmMessages.DELETE_COMPLETED_TODOS,
+    });
   };
 
   return (
@@ -50,7 +53,7 @@ export const TodoListHeader = () => {
           Clear Completed
         </Button>
       </div>
-      <FilteredTodo />
+      <TodoFiltered />
     </div>
   );
 };
